deliverByNotify: Use 64 bits for time variables
This patch addresses the following Coverity report:

CID 380298 (#1 of 1): Use of 32-bit time_t (Y2K38_SAFETY)1. store_truncates_time_t: A time_t value is stored in an integer with too few bits to accommodate it. The expression time(NULL) is cast to int.

+++ agent/mibgroup/deliver/deliverByNotify.c
@@ -457,7 +457,7 @@ _schedule_next_execute_time(void) {
     for(obj = ITERATOR_FIRST(iterator); obj;
         obj = ITERATOR_NEXT(iterator)) {
         next_time = calculate_time_until_next_run(obj, &local_now);
-        DEBUGMSGTL(("deliverByNotify", "  obj: %d (last=%d, next_run=%d)\n", next_time, obj->last_run, obj->next_run));
+        DEBUGMSGTL(("deliverByNotify", "  obj: %d (last=%lld, next_run=%lld)\n", next_time, (long long)obj->last_run, (long long)obj->next_run));
         if (next_time < sleep_for)
             sleep_for = next_time;
     }
